requirements.txt
torch
torchvision
transformers
open_clip_torch
sentencepiece
protobuf
openai
pillow
numpy
pandas
scikit-learn
opencv-python
PyYAML
typer[toml]
tqdm
tabulate
